# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /services/{name}:
    get:
      tags:
        - services
      summary: Returns a registred service by name
      description: Returns a registred service by name
      operationId: getService
      parameters:
        - name: name
          in: path
          description: Name of the service
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDTO'
...
components:
  schemas:
    ServiceDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        port:
          type: string
        registerTime:
          type: string
          format: date-time